requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
bzip2=1.0.8=h516909a_2
ca-certificates=2020.6.24=0
cairo=1.16.0=hcf35c78_1003
certifi=2020.6.20=py38_0
cloudpickle=1.4.1=py_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py38_0
cytoolz=0.10.1=py38h7b6447c_0
dask-core=2.19.0=py_0
dbus=1.13.16=hb2f20db_0
decorator=4.4.2=py_0
easydict=1.9=py_0
expat=2.2.9=he6710b0_2
ffmpeg=4.2.3=h167e202_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.2=h5ab3b9f_0
gettext=0.19.8.1=hc5be6a0_1002
glib=2.65.0=h6f030ca_0
gmp=6.2.0=he1b5a44_2
gnutls=3.6.13=h79a8f9a_0
graphite2=1.3.13=he1b5a44_1001
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
h5py=2.10.0=py38hd6299e0_1
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.6=nompi_h3c11f04_100
icu=64.2=he1b5a44_1
imageio=2.8.0=py_0
intel-openmp=2020.1=217
jasper=1.900.1=h07fcdf6_1006
jpeg=9d=h516909a_0
kiwisolver=1.2.0=py38hfd86e86_0
lame=3.100=h14c3975_1001
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=16_mkl
libcblas=3.8.0=16_mkl
libclang=9.0.1=default_hde54327_0
libedit=3.1.20191231=h7b6447c_0
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1006
liblapack=3.8.0=16_mkl
liblapacke=3.8.0=16_mkl
libllvm9=9.0.1=he513fc3_1
libopencv=4.3.0=py38_1
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.14=h7b6447c_0
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
lz4-c=1.9.2=he6710b0_0
matplotlib=3.2.2=0
matplotlib-base=3.2.2=py38h2af1d28_0
mkl=2020.1=217
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.1.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=h1bed415_1002
networkx=2.4=py_0
ninja=1.9.0=py38hfd86e86_0
nspr=4.26=he1b5a44_0
nss=3.47=he751ad9_0
numpy=1.18.5=py38ha1c710e_0
numpy-base=1.18.5=py38hde5b4d6_0
olefile=0.46=py_0
opencv=4.3.0=py38_1
openh264=2.1.1=h8b12597_0
openssl=1.1.1g=h7b6447c_0
pcre=8.44=he6710b0_0
pillow=7.1.2=py38hb39fc2d_0
pip=20.1.1=py38_1
pixman=0.38.0=h516909a_1003
py-opencv=4.3.0=py38h23f93f0_1
pyparsing=2.4.7=py_0
pyqt=5.12.3=py38ha8c2ead_3
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
python=3.8.3=cpython_he5300dc_0
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.5.1=py3.8_cuda10.2.89_cudnn7.6.5_0
pywavelets=1.1.1=py38h7b6447c_0
pyyaml=5.3.1=py38h7b6447c_1
qt=5.12.5=hd8c4c69_1
readline=8.0=h7b6447c_0
scikit-image=0.16.2=py38h0573a6f_0
scipy=1.5.0=py38h0b6359f_0
setuptools=47.3.1=py38_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py_0
sqlite=3.32.3=h62c20be_0
termcolor=1.1.0=py38_1
tk=8.6.10=hbc83047_0
toolz=0.10.0=py_0
torchvision=0.6.1=py38_cu102
tornado=6.0.4=py38h7b6447c_1
tqdm=4.47.0=py_0
wheel=0.34.2=py38_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
